requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
python-multipart>=0.0.9